Move integration tests into trellis-test module
Resolves #64
