+++ build.gradle
@@ -92,7 +92,7 @@ subprojects {
         apiguardianVersion = '1.0.0'
         awaitilityVersion = '3.0.0'
         commonsTextVersion = '1.2'
-        jerseyVersion = '2.26'
+        jerseyVersion = '2.25.1'
         jsonVersion = '1.1.2'
         junitVersion = '5.1.0'
         junitPlatformVersion = '1.1.0'
@@ -293,7 +293,7 @@ task jacocoRootReport(type: org.gradle.testing.jacoco.tasks.JacocoReport) {
     additionalSourceDirs = files(subprojects.sourceSets.main.allSource.srcDirs)
     sourceDirectories = files(subprojects.sourceSets.main.allSource.srcDirs)
     classDirectories =  files(subprojects.sourceSets.main.output)
-    executionData = files(subprojects.findAll { it.name != 'trellis-karaf' }.jacocoTestReport.executionData)
+    executionData = files(subprojects.findAll { it.name != 'trellis-karaf' && it.name != 'trellis-test' }.jacocoTestReport.executionData)
     reports {
         html.enabled = true
         xml.enabled = true
