+++ trellis-test/src/main/java/org/trellisldp/test/AuthorizationTests.java
@@ -11,15 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.app;
+package org.trellisldp.test;
 
-import static io.dropwizard.testing.ConfigOverride.config;
-import static io.dropwizard.testing.ResourceHelpers.resourceFilePath;
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static java.util.stream.Collectors.toList;
 import static javax.ws.rs.client.Entity.entity;
 import static javax.ws.rs.core.HttpHeaders.AUTHORIZATION;
 import static javax.ws.rs.core.HttpHeaders.LINK;
+import static javax.ws.rs.core.Link.TYPE;
 import static javax.ws.rs.core.Link.fromUri;
 import static javax.ws.rs.core.Response.Status.FORBIDDEN;
 import static javax.ws.rs.core.Response.Status.Family.SUCCESSFUL;
@@ -31,79 +29,61 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.trellisldp.http.domain.RdfMediaType.APPLICATION_SPARQL_UPDATE;
 import static org.trellisldp.http.domain.RdfMediaType.TEXT_TURTLE;
 
-import io.dropwizard.client.JerseyClientBuilder;
-import io.dropwizard.testing.DropwizardTestSupport;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 
-import java.util.List;
-
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.WebTarget;
-import javax.ws.rs.core.Link;
 import javax.ws.rs.core.Response;
 
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
-import org.trellisldp.app.config.TrellisConfiguration;
 import org.trellisldp.vocabulary.LDP;
 import org.trellisldp.vocabulary.Trellis;
 
 /**
- * Authorization tests
+ * Authorization tests.
  *
  * @author acoburn
  */
 @RunWith(JUnitPlatform.class)
-public class TrellisAuthenticationTest {
+public class AuthorizationTests extends BaseCommonTests {
 
     private static final String JWT_SECRET = "secret";
+    private static final String ACL = "acl";
+    private static final String PATCH = "PATCH";
+    private static final String EXT_ACL = "?ext=acl";
+    private static final String INSERT_PROP_FOO = "INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}";
+    private static final String INSERT_PROP_BAR = "INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}";
+    private static final String PREFIX_ACL = "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n";
 
-    private static final DropwizardTestSupport<TrellisConfiguration> APP
-        = new DropwizardTestSupport<TrellisConfiguration>(TrellisApplication.class,
-                resourceFilePath("trellis-config.yml"),
-                config("auth.basic.usersFile", resourceFilePath("users.auth")),
-                config("binaries", resourceFilePath("data") + "/binaries"),
-                config("mementos", resourceFilePath("data") + "/mementos"),
-                config("namespaces", resourceFilePath("data/namespaces.json")));
-
-    private static Client client;
-    private static String baseURL;
     private static String container;
-    private static String publicContainer, publicContainerAcl, publicContainerChild;
-    private static String protectedContainer, protectedContainerAcl, protectedContainerChild;
-    private static String privateContainer, privateContainerAcl, privateContainerChild;
-    private static String groupContainer, groupContainerAcl, groupContainerChild;
-    private static String defaultContainer, defaultContainerAcl, defaultContainerChild;
+    private static String publicContainer;
+    private static String publicContainerAcl;
+    private static String publicContainerChild;
+    private static String protectedContainer;
+    private static String protectedContainerAcl;
+    private static String protectedContainerChild;
+    private static String privateContainer;
+    private static String privateContainerAcl;
+    private static String privateContainerChild;
+    private static String groupContainer;
+    private static String groupContainerAcl;
+    private static String groupContainerChild;
+    private static String defaultContainer;
+    private static String defaultContainerAcl;
+    private static String defaultContainerChild;
     private static String groupResource;
 
-    @BeforeAll
-    public static void setUp() {
-        APP.before();
-        client = new JerseyClientBuilder(APP.getEnvironment()).build("test client");
-        client.property("jersey.config.client.connectTimeout", 5000);
-        client.property("jersey.config.client.readTimeout", 5000);
-        baseURL = "http://localhost:" + APP.getLocalPort() + "/";
+    protected static void setUp() {
+        final String jwt = buildJwt(Trellis.AdministratorAgent.getIRIString(), JWT_SECRET);
 
-        final String jwt = "Bearer " + Jwts.builder().claim("webid",
-                Trellis.AdministratorAgent.getIRIString())
-            .signWith(SignatureAlgorithm.HS512, JWT_SECRET.getBytes(UTF_8))
-            .compact();
-
-        final String containerContent
-            = "PREFIX skos: <http://www.w3.org/2004/02/skos/core#> \n"
-            + "PREFIX dc: <http://purl.org/dc/terms/> \n\n"
-            + "<> skos:prefLabel \"Basic Container\"@eng ; "
-            + "   dc:description \"This is a simple Basic Container for testing.\"@eng .";
+        final String containerContent = getResourceAsString("/basicContainer.ttl");
 
         // POST an LDP-BC
         try (final Response res = target().request()
-                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel("type").build())
+                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel(TYPE).build())
                 .header(AUTHORIZATION, jwt).post(entity(containerContent, TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             container = res.getLocation().toString();
@@ -111,7 +91,7 @@ public class TrellisAuthenticationTest {
 
         // POST a public container
         try (final Response res = target(container).request()
-                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel("type").build())
+                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel(TYPE).build())
                 .header(AUTHORIZATION, jwt).post(entity(containerContent, TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             publicContainer = res.getLocation().toString();
@@ -122,27 +102,27 @@ public class TrellisAuthenticationTest {
                 .post(entity("", TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             publicContainerChild = res.getLocation().toString();
-            publicContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals("acl"))
+            publicContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals(ACL))
                 .map(link -> link.getUri().toString()).findFirst().orElse("");
         }
 
-        final String publicAcl = "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n"
-            + "PREFIX foaf: <http://xmlns.com/foaf/0.1/>\n\n"
-            + "INSERT DATA { \n"
-            + "[acl:accessTo <" + publicContainer + ">; acl:mode acl:Read; acl:agentClass foaf:Agent ] }; \n"
-            + "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n"
-            + "INSERT DATA { [acl:accessTo <" + publicContainer + ">; acl:mode acl:Read, acl:Write; "
+        final String publicAcl = "PREFIX foaf: <http://xmlns.com/foaf/0.1/>\n"
+            + PREFIX_ACL
+            + "INSERT DATA { [acl:accessTo <" + publicContainer + ">; acl:mode acl:Read; "
+            + "   acl:agentClass foaf:Agent ] }; \n"
+            + PREFIX_ACL
+            + "INSERT DATA { [acl:accessTo <" + publicContainer + ">; acl:mode acl:Read, acl:Write;"
             + "   acl:agentClass acl:AuthenticatedAgent ] }";
 
         // Add an ACL for the public container
-        try (final Response res = target(publicContainerAcl).request().header(AUTHORIZATION, jwt).method("PATCH",
+        try (final Response res = target(publicContainerAcl).request().header(AUTHORIZATION, jwt).method(PATCH,
                     entity(publicAcl, APPLICATION_SPARQL_UPDATE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
         }
 
         // POST a protected container
         try (final Response res = target(container).request()
-                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel("type").build())
+                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel(TYPE).build())
                 .header(AUTHORIZATION, jwt).post(entity(containerContent, TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             protectedContainer = res.getLocation().toString();
@@ -153,28 +133,28 @@ public class TrellisAuthenticationTest {
                 .post(entity("", TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             protectedContainerChild = res.getLocation().toString();
-            protectedContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals("acl"))
+            protectedContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals(ACL))
                 .map(link -> link.getUri().toString()).findFirst().orElse("");
         }
 
-        final String protectedAcl = "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n"
+        final String protectedAcl = PREFIX_ACL
             + "INSERT DATA { \n"
-            + "[acl:accessTo <" + protectedContainer + ">; acl:mode acl:Read, acl:Write; "
+            + "[acl:accessTo  <" + protectedContainer + ">;  acl:mode acl:Read, acl:Write;"
             + "   acl:agent <https://people.apache.org/~acoburn/#i> ] };"
-            + "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n"
+            + PREFIX_ACL
             + "INSERT DATA { \n"
-            + "[acl:accessTo <" + protectedContainer + ">; acl:mode acl:Read, acl:Append; "
+            + "[acl:accessTo  <" + protectedContainer + ">; acl:mode acl:Read, acl:Append; "
             + "   acl:agent <https://madison.example.com/profile/#me> ] }";
 
         // Add an ACL for the protected container
-        try (final Response res = target(protectedContainerAcl).request().header(AUTHORIZATION, jwt).method("PATCH",
+        try (final Response res = target(protectedContainerAcl).request().header(AUTHORIZATION, jwt).method(PATCH,
                     entity(protectedAcl, APPLICATION_SPARQL_UPDATE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
         }
 
         // POST a private container
         try (final Response res = target(container).request()
-                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel("type").build())
+                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel(TYPE).build())
                 .header(AUTHORIZATION, jwt).post(entity(containerContent, TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             privateContainer = res.getLocation().toString();
@@ -185,17 +165,17 @@ public class TrellisAuthenticationTest {
                 .post(entity("", TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             privateContainerChild = res.getLocation().toString();
-            privateContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals("acl"))
+            privateContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals(ACL))
                 .map(link -> link.getUri().toString()).findFirst().orElse("");
         }
 
-        final String privateAcl = "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n"
+        final String privateAcl = PREFIX_ACL
             + "INSERT DATA { "
-            + "[acl:accessTo <" + privateContainer + ">; acl:mode acl:Read, acl:Write; "
+            + "[acl:accessTo  <" + privateContainer + ">; acl:mode acl:Read, acl:Write; "
             + "   acl:agent <http://example.com/administrator> ] }";
 
         // Add an ACL for the private container
-        try (final Response res = target(privateContainerAcl).request().header(AUTHORIZATION, jwt).method("PATCH",
+        try (final Response res = target(privateContainerAcl).request().header(AUTHORIZATION, jwt).method(PATCH,
                     entity(privateAcl, APPLICATION_SPARQL_UPDATE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
         }
@@ -221,7 +201,7 @@ public class TrellisAuthenticationTest {
 
         // POST a group-controlled container
         try (final Response res = target(container).request()
-                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel("type").build())
+                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel(TYPE).build())
                 .header(AUTHORIZATION, jwt).post(entity(containerContent, TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             groupContainer = res.getLocation().toString();
@@ -232,28 +212,28 @@ public class TrellisAuthenticationTest {
                 .post(entity("", TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             groupContainerChild = res.getLocation().toString();
-            groupContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals("acl"))
+            groupContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals(ACL))
                 .map(link -> link.getUri().toString()).findFirst().orElse("");
         }
 
-        final String groupAcl = "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n"
-            + "INSERT DATA { "
+        final String groupAcl = PREFIX_ACL
+            + "INSERT DATA {  "
             + "[acl:accessTo <" + groupContainer + ">; acl:mode acl:Read, acl:Write; "
-            + "   acl:agentGroup <" + groupResource + "#Developers> ] };\n"
-            + "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n"
-            + "INSERT DATA { "
+            + " acl:agentGroup <" + groupResource + "#Developers> ] };\n"
+            + PREFIX_ACL
+            + "INSERT DATA {  "
             + "[acl:accessTo <" + groupContainer + ">; acl:mode acl:Read; "
-            + "   acl:agentGroup <" + groupResource + "#Management> ] }";
+            + " acl:agentGroup <" + groupResource + "#Management> ] }";
 
         // Add an ACL for the private container
-        try (final Response res = target(groupContainerAcl).request().header(AUTHORIZATION, jwt).method("PATCH",
+        try (final Response res = target(groupContainerAcl).request().header(AUTHORIZATION, jwt).method(PATCH,
                     entity(groupAcl, APPLICATION_SPARQL_UPDATE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
         }
 
         // POST a container with a default ACL
         try (final Response res = target(container).request()
-                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel("type").build())
+                .header(LINK, fromUri(LDP.BasicContainer.getIRIString()).rel(TYPE).build())
                 .header(AUTHORIZATION, jwt).post(entity(containerContent, TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             defaultContainer = res.getLocation().toString();
@@ -264,46 +244,53 @@ public class TrellisAuthenticationTest {
                 .post(entity("", TEXT_TURTLE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             defaultContainerChild = res.getLocation().toString();
-            defaultContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals("acl"))
+            defaultContainerAcl = getLinks(res).stream().filter(link -> link.getRel().equals(ACL))
                 .map(link -> link.getUri().toString()).findFirst().orElse("");
         }
 
-        final String defaultAcl = "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n"
-            + "PREFIX foaf: <http://xmlns.com/foaf/0.1/>\n\n"
-            + "INSERT DATA { \n"
+        final String defaultAcl = "PREFIX foaf: <http://xmlns.com/foaf/0.1/>\n"
+            + PREFIX_ACL
+            + "INSERT DATA {  "
             + "[acl:accessTo <" + defaultContainer + ">; acl:mode acl:Read; acl:agentClass foaf:Agent ] }; \n"
-            + "PREFIX acl: <http://www.w3.org/ns/auth/acl#>\n\n"
+            + PREFIX_ACL
             + "INSERT DATA { [acl:accessTo <" + defaultContainer + ">; acl:mode acl:Read, acl:Write; \n"
             + "   acl:default <" + defaultContainer + ">; \n"
             + "   acl:agent <https://people.apache.org/~acoburn/#i> ] }";
 
         // Add an ACL for the public container
-        try (final Response res = target(defaultContainerAcl).request().header(AUTHORIZATION, jwt).method("PATCH",
+        try (final Response res = target(defaultContainerAcl).request().header(AUTHORIZATION, jwt).method(PATCH,
                     entity(defaultAcl, APPLICATION_SPARQL_UPDATE))) {
             assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
         }
     }
 
-    @AfterAll
-    public static void tearDown() {
-        APP.after();
-    }
-
+    /**
+     * Test for an ACL Link header on a public resource.
+     */
     @Test
     public void testPublicLinkHeader() {
-        assertEquals(publicContainer + "?ext=acl", publicContainerAcl);
+        assertEquals(publicContainer + EXT_ACL, publicContainerAcl);
     }
 
+    /**
+     * Test for an ACL Link header on a protected resource.
+     */
     @Test
     public void testProtectedLinkHeader() {
-        assertEquals(protectedContainer + "?ext=acl", protectedContainerAcl);
+        assertEquals(protectedContainer + EXT_ACL, protectedContainerAcl);
     }
 
+    /**
+     * Test for an ACL Link header on a private resource.
+     */
     @Test
     public void testPrivateLinkHeader() {
-        assertEquals(privateContainer + "?ext=acl", privateContainerAcl);
+        assertEquals(privateContainer + EXT_ACL, privateContainerAcl);
     }
 
+    /**
+     * Jwt Administrator tests.
+     */
     @Nested
     @DisplayName("Jwt Administrator tests")
     public class JwtAdministratorTests {
@@ -313,6 +300,9 @@ public class TrellisAuthenticationTest {
             .signWith(SignatureAlgorithm.HS512, JWT_SECRET.getBytes(UTF_8))
             .compact();
 
+        /**
+         * Verify that an administrator can read a public resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read a public resource")
         public void testAdminCanReadPublicResource() {
@@ -322,6 +312,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can read the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read the child of a public resource")
         public void testAdminCanReadPublicResourceChild() {
@@ -331,28 +324,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can write to a public resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to a public resource")
         public void testAdminCanWritePublicResource() {
             try (final Response res = target(publicContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can write to the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to the child of a public resource")
         public void testAdminCanWritePublicResourceChild() {
             try (final Response res = target(publicContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can control a public resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control a public resource")
         public void testAdminCanControlPublicResource() {
@@ -362,15 +360,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can control the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control the child of a public resource")
         public void testAdminCanControlPublicResourceChild() {
-            try (final Response res = target(publicContainerChild + "?ext=acl").request()
+            try (final Response res = target(publicContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(NOT_FOUND, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an administrator can read a protected resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read a protected resource")
         public void testAdminCanReadProtectedResource() {
@@ -380,6 +384,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can read the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read the child of a protected resource")
         public void testAdminCanReadProtectedResourceChild() {
@@ -389,28 +396,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can write to a protected resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to a protected resource")
         public void testAdminCanWriteProtectedResource() {
             try (final Response res = target(protectedContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can write to the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to the child of a protected resource")
         public void testAdminCanWriteProtectedResourceChild() {
             try (final Response res = target(protectedContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can control a protected resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control a protected resource")
         public void testAdminCanControlProtectedResource() {
@@ -420,15 +432,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can control the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control the child of a protected resource")
         public void testAdminCanControlProtectedResourceChild() {
-            try (final Response res = target(protectedContainerChild + "?ext=acl").request()
+            try (final Response res = target(protectedContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(NOT_FOUND, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an administrator can read a private resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read a private resource")
         public void testAdminCanReadPrivateResource() {
@@ -438,6 +456,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can read the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read the child of a private resource")
         public void testAdminCanReadPrivateResourceChild() {
@@ -447,28 +468,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can write to a private resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to a private resource")
         public void testAdminCanWritePrivateResource() {
             try (final Response res = target(privateContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can write to the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to the child of a private resource")
         public void testAdminCanWritePrivateResourceChild() {
             try (final Response res = target(privateContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can control a private resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control a private resource")
         public void testAdminCanControlPrivateResource() {
@@ -478,15 +504,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can control the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control the child of a private resource")
         public void testAdminCanControlPrivateResourceChild() {
-            try (final Response res = target(privateContainerChild + "?ext=acl").request()
+            try (final Response res = target(privateContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(NOT_FOUND, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an administrator can read a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read a group-controlled resource")
         public void testAdminCanReadGroupResource() {
@@ -496,6 +528,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can read the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read the child of a group-controlled resource")
         public void testAdminCanReadGroupResourceChild() {
@@ -505,28 +540,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can write to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to a group-controlled resource")
         public void testAdminCanWriteGroupResource() {
             try (final Response res = target(groupContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can write to the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to the child of a group-controlled resource")
         public void testAdminCanWriteGroupResourceChild() {
             try (final Response res = target(groupContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can control a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control a group-controlled resource")
         public void testAdminCanControlGroupResource() {
@@ -536,15 +576,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can't find the ACL of a child resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can't find the ACL of a child resource")
         public void testAdminCanControlGroupResourceChild() {
-            try (final Response res = target(groupContainerChild + "?ext=acl").request()
+            try (final Response res = target(groupContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(NOT_FOUND, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an administrator can read a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read a default ACL resource")
         public void testCanReadDefaultAclResource() {
@@ -554,6 +600,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can read the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can read the child of a default ACL resource")
         public void testCanReadDefaultAclResourceChild() {
@@ -563,28 +612,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can write to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to a default ACL resource")
         public void testCanWriteDefaultAclResource() {
             try (final Response res = target(defaultContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can write to the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can write to the child of a default ACL resource")
         public void testCanWriteDefaultAclResourceChild() {
             try (final Response res = target(defaultContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that an administrator can control a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can control a default ACL resource")
         public void testCanControlDefaultAclResource() {
@@ -594,22 +648,31 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an administrator can't find the ACL resource.
+         */
         @Test
         @DisplayName("Verify that an administrator can't find the ACL resource")
         public void testCanControlDefaultAclResourceChild() {
-            try (final Response res = target(defaultContainerChild + "?ext=acl").request()
+            try (final Response res = target(defaultContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(NOT_FOUND, fromStatusCode(res.getStatus()));
             }
         }
     }
 
+    /**
+     * Basic Auth User tests.
+     */
     @Nested
     @DisplayName("Basic Auth User tests")
     public class BasicAuthUserTests {
 
         private final String auth = "Basic " + encodeBase64String("acoburn:secret".getBytes());
 
+        /**
+         * Verify that a user can read a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a public resource")
         public void testUserCanReadPublicResource() {
@@ -619,6 +682,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a public resource")
         public void testUserCanReadPublicResourceChild() {
@@ -628,6 +694,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a public resource")
         public void testUserCanAppendPublicResource() {
@@ -637,46 +706,57 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a public resource")
         public void testUserCanWritePublicResource() {
             try (final Response res = target(publicContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a public resource")
         public void testUserCanWritePublicResourceChild() {
             try (final Response res = target(publicContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a public resource")
         public void testUserCanControlPublicResource() {
-            try (final Response res = target(publicContainer + "?ext=acl").request()
+            try (final Response res = target(publicContainer + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a public resource")
         public void testUserCanControlPublicResourceChild() {
-            try (final Response res = target(publicContainerChild + "?ext=acl").request()
+            try (final Response res = target(publicContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a protected resource")
         public void testUserCanReadProtectedResource() {
@@ -686,6 +766,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a protected resource")
         public void testUserCanReadProtectedResourceChild() {
@@ -695,6 +778,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a protected resource")
         public void testUserCanAppendProtectedResource() {
@@ -704,28 +790,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a protected resource")
         public void testUserCanWriteProtectedResource() {
             try (final Response res = target(protectedContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a protected resource")
         public void testUserCanWriteProtectedResourceChild() {
             try (final Response res = target(protectedContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a protected resource")
         public void testUserCanControlProtectedResource() {
@@ -735,15 +826,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a protected resource")
         public void testUserCanControlProtectedResourceChild() {
-            try (final Response res = target(protectedContainerChild + "?ext=acl").request()
+            try (final Response res = target(protectedContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot read a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read a private resource")
         public void testUserCanReadPrivateResource() {
@@ -753,6 +850,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot read the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read the child of a private resource")
         public void testUserCanReadPrivateResourceChild() {
@@ -762,6 +862,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a private resource")
         public void testUserCanAppendPrivateResource() {
@@ -771,46 +874,57 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a private resource")
         public void testUserCanWritePrivateResource() {
             try (final Response res = target(privateContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a private resource")
         public void testUserCanWritePrivateResourceChild() {
             try (final Response res = target(privateContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a private resource")
         public void testUserCanControlPrivateResource() {
-            try (final Response res = target(privateContainer + "?ext=acl").request()
+            try (final Response res = target(privateContainer + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a private resource")
         public void testUserCanControlPrivateResourceChild() {
-            try (final Response res = target(privateContainerChild + "?ext=acl").request()
+            try (final Response res = target(privateContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a group-controlled resource")
         public void testCanReadGroupResource() {
@@ -820,6 +934,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a group-controlled resource")
         public void testCanReadGroupResourceChild() {
@@ -829,28 +946,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a group-controlled resource")
         public void testCanWriteGroupResource() {
             try (final Response res = target(groupContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a group-controlled resource")
         public void testCanWriteGroupResourceChild() {
             try (final Response res = target(groupContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a group-controlled resource")
         public void testCanControlGroupResource() {
@@ -860,15 +982,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a group-controlled resource")
         public void testCanControlGroupResourceChild() {
-            try (final Response res = target(groupContainerChild + "?ext=acl").request()
+            try (final Response res = target(groupContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a default ACL resource")
         public void testCanReadDefaultAclResource() {
@@ -878,6 +1006,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a default ACL resource")
         public void testCanReadDefaultAclResourceChild() {
@@ -887,28 +1018,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a default ACL resource")
         public void testCanWriteDefaultAclResource() {
             try (final Response res = target(defaultContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a default ACL resource")
         public void testCanWriteDefaultAclResourceChild() {
             try (final Response res = target(defaultContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a default ACL resource")
         public void testCanControlDefaultAclResource() {
@@ -918,25 +1054,31 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a default ACL resource")
         public void testCanControlDefaultAclResourceChild() {
-            try (final Response res = target(defaultContainerChild + "?ext=acl").request()
+            try (final Response res = target(defaultContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
     }
 
+    /**
+     * JWT user tests.
+     */
     @Nested
     @DisplayName("Jwt User tests")
     public class JwtUserTests {
 
-        private final String jwt = "Bearer " + Jwts.builder()
-            .claim("webid", "https://people.apache.org/~acoburn/#i")
-            .signWith(SignatureAlgorithm.HS512, JWT_SECRET.getBytes(UTF_8))
-            .compact();
+        private final String jwt = buildJwt("https://people.apache.org/~acoburn/#i", JWT_SECRET);
 
+        /**
+         * Verify that a user can read a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a public resource")
         public void testUserCanReadPublicResource() {
@@ -946,6 +1088,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a public resource")
         public void testUserCanReadPublicResourceChild() {
@@ -955,6 +1100,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a public resource")
         public void testUserCanAppendPublicResource() {
@@ -964,28 +1112,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a public resource")
         public void testUserCanWritePublicResource() {
             try (final Response res = target(publicContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a public resource")
         public void testUserCanWritePublicResourceChild() {
             try (final Response res = target(publicContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a public resource")
         public void testUserCanControlPublicResource() {
@@ -995,15 +1148,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a public resource")
         public void testUserCanControlPublicResourceChild() {
-            try (final Response res = target(publicContainerChild + "?ext=acl").request()
+            try (final Response res = target(publicContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a protected resource")
         public void testUserCanReadProtectedResource() {
@@ -1013,6 +1172,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a protected resource")
         public void testUserCanReadProtectedResourceChild() {
@@ -1022,6 +1184,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a protected resource")
         public void testUserCanAppendProtectedResource() {
@@ -1031,28 +1196,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a protected resource")
         public void testUserCanWriteProtectedResource() {
             try (final Response res = target(protectedContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a protected resource")
         public void testUserCanWriteProtectedResourceChile() {
             try (final Response res = target(protectedContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a protected resource")
         public void testUserCanControlProtectedResource() {
@@ -1062,15 +1232,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a protected resource")
         public void testUserCanControlProtectedResourceChild() {
-            try (final Response res = target(protectedContainerChild + "?ext=acl").request()
+            try (final Response res = target(protectedContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot read a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read a private resource")
         public void testUserCanReadPrivateResource() {
@@ -1080,6 +1256,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot read the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read the child of a private resource")
         public void testUserCanReadPrivateResourceChile() {
@@ -1089,6 +1268,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a private resource")
         public void testUserCanAppendPrivateResource() {
@@ -1098,28 +1280,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a private resource")
         public void testUserCanWritePrivateResource() {
             try (final Response res = target(privateContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a private resource")
         public void testUserCanWritePrivateResourceChild() {
             try (final Response res = target(privateContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a private resource")
         public void testUserCanControlPrivateResource() {
@@ -1129,15 +1316,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a private resource")
         public void testUserCanControlPrivateResourceChild() {
-            try (final Response res = target(privateContainerChild + "?ext=acl").request()
+            try (final Response res = target(privateContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a group-controlled resource")
         public void testCanReadGroupResource() {
@@ -1147,6 +1340,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a group-controlled resource")
         public void testCanReadGroupResourceChild() {
@@ -1156,28 +1352,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a group-controlled resource")
         public void testCanWriteGroupResource() {
             try (final Response res = target(groupContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a group-controlled resource")
         public void testCanWriteGroupResourceChild() {
             try (final Response res = target(groupContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a group-controlled resource")
         public void testCanControlGroupResource() {
@@ -1187,15 +1388,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a group-controlled resource")
         public void testCanControlGroupResourceChild() {
-            try (final Response res = target(groupContainerChild + "?ext=acl").request()
+            try (final Response res = target(groupContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a default ACL resource")
         public void testCanReadDefaultAclResource() {
@@ -1205,6 +1412,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a default ACL resource")
         public void testCanReadDefaultAclResourceChild() {
@@ -1214,28 +1424,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a default ACL resource")
         public void testCanWriteDefaultAclResource() {
             try (final Response res = target(defaultContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a default ACL resource")
         public void testCanWriteDefaultAclResourceChild() {
             try (final Response res = target(defaultContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a default ACL resource")
         public void testCanControlDefaultAclResource() {
@@ -1245,22 +1460,31 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a default ACL resource")
         public void testCanControlDefaultAclResourceChild() {
-            try (final Response res = target(defaultContainerChild + "?ext=acl").request()
+            try (final Response res = target(defaultContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
     }
 
+    /**
+     * Basic Auth other user tests.
+     */
     @Nested
     @DisplayName("Basic Auth other user tests")
     public class BasicAuthOtherUserTests {
 
         private final String auth = "Basic " + encodeBase64String("user:password".getBytes(UTF_8));
 
+        /**
+         * Verify that a user can read a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a public resource")
         public void testUserCanReadPublicResource() {
@@ -1270,6 +1494,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a public resource")
         public void testUserCanReadPublicResourceChile() {
@@ -1279,6 +1506,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a public resource")
         public void testUserCanAppendPublicResource() {
@@ -1288,46 +1518,57 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a public resource")
         public void testUserCanWritePublicResource() {
             try (final Response res = target(publicContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a public resource")
         public void testUserCanWritePublicResourceChild() {
             try (final Response res = target(publicContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a public resource")
         public void testUserCanControlPublicResource() {
-            try (final Response res = target(publicContainer + "?ext=acl").request()
+            try (final Response res = target(publicContainer + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a public resource")
         public void testUserCanControlPublicResourceChild() {
-            try (final Response res = target(publicContainerChild + "?ext=acl").request()
+            try (final Response res = target(publicContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a protected resource")
         public void testUserCanReadProtectedResource() {
@@ -1337,6 +1578,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a protected resource")
         public void testUserCanReadProtectedResourceChild() {
@@ -1346,6 +1590,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a public resource")
         public void testUserCanAppendProtectedResource() {
@@ -1355,28 +1602,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a protected resource")
         public void testUserCanWriteProtectedResource() {
             try (final Response res = target(protectedContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a protected resource")
         public void testUserCanWriteProtectedResourceChild() {
             try (final Response res = target(protectedContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a protected resource")
         public void testUserCanControlProtectedResource() {
@@ -1386,15 +1638,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a protected resource")
         public void testUserCanControlProtectedResourceChild() {
-            try (final Response res = target(protectedContainerChild + "?ext=acl").request()
+            try (final Response res = target(protectedContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot read a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read a private resource")
         public void testUserCanReadPrivateResource() {
@@ -1404,6 +1662,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot read the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read the child of a private resource")
         public void testUserCanReadPrivateResourceChild() {
@@ -1413,6 +1674,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a private resource")
         public void testUserCanAppendPrivateResource() {
@@ -1422,28 +1686,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a private resource")
         public void testUserCanWritePrivateResource() {
             try (final Response res = target(privateContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a private resource")
         public void testUserCanWritePrivateResourceChild() {
             try (final Response res = target(privateContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a private resource")
         public void testUserCanControlPrivateResource() {
@@ -1453,15 +1722,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a private resource")
         public void testUserCanControlPrivateResourceChild() {
-            try (final Response res = target(privateContainerChild + "?ext=acl").request()
+            try (final Response res = target(privateContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a group-controlled resource")
         public void testCanReadGroupResource() {
@@ -1471,6 +1746,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a group-controlled resource")
         public void testCanReadGroupResourceChild() {
@@ -1480,6 +1758,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a group-controlled resource")
         public void testUserCanAppendGroupResource() {
@@ -1489,28 +1770,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a group-controlled resource")
         public void testCanWriteGroupResource() {
             try (final Response res = target(groupContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a group-controlled resource")
         public void testCanWriteGroupResourceChild() {
             try (final Response res = target(groupContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a group-controlled resource")
         public void testCanControlGroupResource() {
@@ -1520,15 +1806,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a group-controlled resource")
         public void testCanControlGroupResourceChild() {
-            try (final Response res = target(groupContainerChild + "?ext=acl").request()
+            try (final Response res = target(groupContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a default ACL resource")
         public void testCanReadDefaultAclResource() {
@@ -1538,6 +1830,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot read the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read the child of a default ACL resource")
         public void testCanReadDefaultAclResourceChild() {
@@ -1547,6 +1842,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a default ACL resource")
         public void testUserCanAppendDefaultAclResource() {
@@ -1556,28 +1854,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a default ACL resource")
         public void testCanWriteDefaultAclResource() {
             try (final Response res = target(defaultContainer).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a default ACL resource")
         public void testCanWriteDefaultAclResourceChild() {
             try (final Response res = target(defaultContainerChild).request()
-                    .header(AUTHORIZATION, auth).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, auth).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a default ACL resource")
         public void testCanControlDefaultAclResource() {
@@ -1587,25 +1890,31 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a default ACL resource")
         public void testCanControlDefaultAclResourceChild() {
-            try (final Response res = target(defaultContainerChild + "?ext=acl").request()
+            try (final Response res = target(defaultContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, auth).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
     }
 
+    /**
+     * JWT Other user tests.
+     */
     @Nested
     @DisplayName("Jwt Other user tests")
     public class JwtOtherUserTests {
 
-        private final String jwt = "Bearer " + Jwts.builder()
-            .claim("webid", "https://madison.example.com/profile/#me")
-            .signWith(SignatureAlgorithm.HS512, JWT_SECRET.getBytes(UTF_8))
-            .compact();
+        private final String jwt = buildJwt("https://madison.example.com/profile/#me", JWT_SECRET);
 
+        /**
+         * Verify that a user can read a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a public resource")
         public void testUserCanReadPublicResource() {
@@ -1615,6 +1924,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a public resource")
         public void testUserCanReadPublicResourceChild() {
@@ -1624,6 +1936,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a public resource")
         public void testUserCanAppendPublicResource() {
@@ -1633,46 +1948,57 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can write to a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to a public resource")
         public void testUserCanWritePublicResource() {
             try (final Response res = target(publicContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user can write to the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user can write to the child of a public resource")
         public void testUserCanWritePublicResourceChild() {
             try (final Response res = target(publicContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(SUCCESSFUL, res.getStatusInfo().getFamily());
             }
         }
 
+        /**
+         * Verify that a user cannot control a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a public resource")
         public void testUserCanControlPublicResource() {
-            try (final Response res = target(publicContainer + "?ext=acl").request()
+            try (final Response res = target(publicContainer + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a public resource")
         public void testUserCanControlPublicResourceChild() {
-            try (final Response res = target(publicContainerChild + "?ext=acl").request()
+            try (final Response res = target(publicContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a protected resource")
         public void testUserCanReadProtectedResource() {
@@ -1682,6 +2008,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a protected resource")
         public void testUserCanReadProtectedResourceChild() {
@@ -1691,6 +2020,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can append to a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user can append to a protected resource")
         public void testUserCanAppendProtectedResource() {
@@ -1700,46 +2032,57 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a protected resource")
         public void testUserCanWriteProtectedResource() {
             try (final Response res = target(protectedContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a protected resource")
         public void testUserCanWriteProtectedResourceChild() {
             try (final Response res = target(protectedContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a protected resource")
         public void testUserCanControlProtectedResource() {
-            try (final Response res = target(protectedContainer + "?ext=acl").request()
+            try (final Response res = target(protectedContainer + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a protected resource")
         public void testUserCanControlProtectedResourceChild() {
-            try (final Response res = target(protectedContainerChild + "?ext=acl").request()
+            try (final Response res = target(protectedContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot read a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read a private resource")
         public void testUserCanReadPrivateResource() {
@@ -1749,6 +2092,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot read the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read the child of a private resource")
         public void testUserCanReadPrivateResourceChild() {
@@ -1758,6 +2104,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a private resource")
         public void testUserCanAppendPrivateResource() {
@@ -1767,28 +2116,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a private resource")
         public void testUserCanWritePrivateResource() {
             try (final Response res = target(privateContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a private resource")
         public void testUserCanWritePrivateResourceChild() {
             try (final Response res = target(privateContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a private resource")
         public void testUserCanControlPrivateResource() {
@@ -1798,15 +2152,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a private resource")
         public void testUserCanControlPrivateResourceChild() {
-            try (final Response res = target(privateContainerChild + "?ext=acl").request()
+            try (final Response res = target(privateContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a group-controlled resource")
         public void testCanReadGroupResource() {
@@ -1816,6 +2176,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user can read the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user can read the child of a group-controlled resource")
         public void testCanReadGroupResourceChild() {
@@ -1825,6 +2188,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a group-controlled resource")
         public void testUserCanAppendGroupResource() {
@@ -1834,28 +2200,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a group-controlled resource")
         public void testCanWriteGroupResource() {
             try (final Response res = target(groupContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a group-controlled resource")
         public void testCanWriteGroupResourceChild() {
             try (final Response res = target(groupContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a group-controlled resource")
         public void testCanControlGroupResource() {
@@ -1865,15 +2236,21 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a group-controlled resource")
         public void testCanControlGroupResourceChild() {
-            try (final Response res = target(groupContainerChild + "?ext=acl").request()
+            try (final Response res = target(groupContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user can read a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user can read a default ACL resource")
         public void testCanReadDefaultAclResource() {
@@ -1883,6 +2260,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot read the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot read the child of a default ACL resource")
         public void testCanReadDefaultAclResourceChild() {
@@ -1892,6 +2272,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot append to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot append to a group-controlled resource")
         public void testUserCanAppendDefaultAclResource() {
@@ -1901,28 +2284,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot write to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to a default ACL resource")
         public void testCanWriteDefaultAclResource() {
             try (final Response res = target(defaultContainer).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot write to the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot write to the child of a default ACL resource")
         public void testCanWriteDefaultAclResourceChild() {
             try (final Response res = target(defaultContainerChild).request()
-                    .header(AUTHORIZATION, jwt).method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+                    .header(AUTHORIZATION, jwt).method(PATCH, entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a default ACL resource")
         public void testCanControlDefaultAclResource() {
@@ -1932,20 +2320,29 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a default ACL resource")
         public void testCanControlDefaultAclResourceChild() {
-            try (final Response res = target(defaultContainerChild + "?ext=acl").request()
+            try (final Response res = target(defaultContainerChild + EXT_ACL).request()
                     .header(AUTHORIZATION, jwt).get()) {
                 assertEquals(FORBIDDEN, fromStatusCode(res.getStatus()));
             }
         }
     }
 
+    /**
+     * Anonymous tests.
+     */
     @Nested
     @DisplayName("Anonymous tests")
     public class AnonymousTests {
 
+        /**
+         * Verify that an anonymous user can read a public resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user can read a public resource")
         public void testCanReadPublicResource() {
@@ -1954,6 +2351,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user can read the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user can read the child of a public resource")
         public void testCanReadPublicResourceChild() {
@@ -1962,6 +2362,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot append to a public resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot append to a public resource")
         public void testUserCanAppendPublicResource() {
@@ -1970,26 +2373,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to a public resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to a public resource")
         public void testCanWritePublicResource() {
-            try (final Response res = target(publicContainer).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
+            try (final Response res = target(publicContainer).request().method(PATCH, entity(INSERT_PROP_BAR,
                             APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to the child of a public resource")
         public void testCanWritePublicResourceChild() {
-            try (final Response res = target(publicContainerChild).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(publicContainerChild).request().method(PATCH,
+                        entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control a public resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control a public resource")
         public void testCanControlPublicResource() {
@@ -1998,14 +2408,20 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control the child of a public resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control the child of a public resource")
         public void testCanControlPublicResourceChild() {
-            try (final Response res = target(publicContainerChild + "?ext=acl").request().get()) {
+            try (final Response res = target(publicContainerChild + EXT_ACL).request().get()) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot read a protected resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot read a protected resource")
         public void testCanReadProtectedResource() {
@@ -2014,6 +2430,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot read the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot read the child of a protected resource")
         public void testCanReadProtectedResourceChild() {
@@ -2022,6 +2441,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot append to a protected resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot append to a protected resource")
         public void testUserCanAppendProtectedResource() {
@@ -2030,26 +2452,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to a protected resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to a protected resource")
         public void testCanWriteProtectedResource() {
-            try (final Response res = target(protectedContainer).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(protectedContainer).request().method(PATCH,
+                        entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to the child of a protected resource")
         public void testCanWriteProtectedResourceChild() {
-            try (final Response res = target(protectedContainerChild).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Bar\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(protectedContainerChild).request().method(PATCH,
+                        entity(INSERT_PROP_BAR, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control a protected resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control a protected resource")
         public void testCanControlProtectedResource() {
@@ -2058,14 +2487,20 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control the child of a protected resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control the child of a protected resource")
         public void testCanControlProtectedResourceChild() {
-            try (final Response res = target(protectedContainerChild + "?ext=acl").request().get()) {
+            try (final Response res = target(protectedContainerChild + EXT_ACL).request().get()) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot read a private resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot read a private resource")
         public void testCanReadPrivateResource() {
@@ -2074,6 +2509,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot read the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot read the child of a private resource")
         public void testCanReadPrivateResourceChild() {
@@ -2082,6 +2520,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot append to a private resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot append to a private resource")
         public void testUserCanAppendPrivateResource() {
@@ -2090,26 +2531,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to a private resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to a private resource")
         public void testCanWritePrivateResource() {
-            try (final Response res = target(privateContainer).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(privateContainer).request().method(PATCH,
+                        entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to the child of a private resource")
         public void testCanWritePrivateResourceChild() {
-            try (final Response res = target(privateContainerChild).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(privateContainerChild).request().method(PATCH,
+                        entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control a private resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control a private resource")
         public void testCanControlPrivateResource() {
@@ -2118,14 +2566,20 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control the child of a private resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control the child of a private resource")
         public void testCanControlPrivateResourceChild() {
-            try (final Response res = target(privateContainerChild + "?ext=acl").request().get()) {
+            try (final Response res = target(privateContainerChild + EXT_ACL).request().get()) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot read a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot read a group-controlled resource")
         public void testCanReadGroupResource() {
@@ -2134,6 +2588,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot read the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot read the child of a group-controlled resource")
         public void testCanReadGroupResourceChild() {
@@ -2142,6 +2599,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot append to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot append to a group-controlled resource")
         public void testUserCanAppendGroupResource() {
@@ -2150,26 +2610,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to a group-controlled resource")
         public void testCanWriteGroupResource() {
-            try (final Response res = target(groupContainer).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(groupContainer).request().method(PATCH,
+                        entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to the child of a group-controlled resource")
         public void testCanWriteGroupResourceChild() {
-            try (final Response res = target(groupContainerChild).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(groupContainerChild).request().method(PATCH,
+                        entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control a group-controlled resource")
         public void testCanControlGroupResource() {
@@ -2178,14 +2645,20 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot control the child of a group-controlled resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot control the child of a group-controlled resource")
         public void testCanControlGroupResourceChild() {
-            try (final Response res = target(groupContainerChild + "?ext=acl").request().get()) {
+            try (final Response res = target(groupContainerChild + EXT_ACL).request().get()) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user can read a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user can read a default ACL resource")
         public void testCanReadDefaultAclResource() {
@@ -2194,6 +2667,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot read the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot read the child of a default ACL resource")
         public void testCanReadDefaultAclResourceChild() {
@@ -2202,6 +2678,9 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot append to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot append to a default ACL resource")
         public void testUserCanAppendDefaultAclResource() {
@@ -2210,26 +2689,33 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to a default ACL resource")
         public void testCanWriteDefaultAclResource() {
-            try (final Response res = target(defaultContainer).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(defaultContainer).request().method(PATCH,
+                        entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that an anonymous user cannot write to the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that an anonymous user cannot write to the child of a default ACL resource")
         public void testCanWriteDefaultAclResourceChild() {
-            try (final Response res = target(defaultContainerChild).request().method("PATCH",
-                        entity("INSERT { <> <http://example.com/prop> \"Foo\" } WHERE {}",
-                            APPLICATION_SPARQL_UPDATE))) {
+            try (final Response res = target(defaultContainerChild).request().method(PATCH,
+                        entity(INSERT_PROP_FOO, APPLICATION_SPARQL_UPDATE))) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
 
+        /**
+         * Verify that a user cannot control a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control a default ACL resource")
         public void testCanControlDefaultAclResource() {
@@ -2238,25 +2724,15 @@ public class TrellisAuthenticationTest {
             }
         }
 
+        /**
+         * Verify that a user cannot control the child of a default ACL resource.
+         */
         @Test
         @DisplayName("Verify that a user cannot control the child of a default ACL resource")
         public void testCanControlDefaultAclResourceChild() {
-            try (final Response res = target(defaultContainerChild + "?ext=acl").request().get()) {
+            try (final Response res = target(defaultContainerChild + EXT_ACL).request().get()) {
                 assertEquals(UNAUTHORIZED, fromStatusCode(res.getStatus()));
             }
         }
-    }
-
-    private static WebTarget target() {
-        return target(baseURL);
-    }
-
-    private static WebTarget target(final String url) {
-        return client.target(url);
-    }
-
-    private static List<Link> getLinks(final Response res) {
-        // Jersey's client doesn't parse complex link headers correctly
-        return res.getStringHeaders().get(LINK).stream().map(Link::valueOf).collect(toList());
     }
 }
