+++ trellis-test/src/main/resources/basicContainer.ttl
@@ -1,0 +1,5 @@
+PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
+PREFIX dc: <http://purl.org/dc/terms/>
+
+<> skos:prefLabel "Basic Container"@eng ;
+   dc:description "This is a simple Basic Container for testing."@eng .
