+++ trellis-test/src/main/resources/childResource.ttl
@@ -1,0 +1,7 @@
+PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
+PREFIX foaf: <http://xmlns.com/foaf/0.1/>
+PREFIX dc: <http://purl.org/dc/terms/>
+
+<> skos:prefLabel "Child Resource"@eng ;
+   foaf:primaryTopic <#it> ;
+   dc:description "This is a simple child resource for testing."@eng .
