+++ trellis-test/src/main/resources/data/namespaces.json
@@ -1,0 +1,18 @@
+{
+    "acl" : "http://www.w3.org/ns/auth/acl#",
+    "as" : "https://www.w3.org/ns/activitystreams#",
+    "dc" : "http://purl.org/dc/terms/",
+    "dc11" : "http://purl.org/dc/elements/1.1/",
+    "geo" : "http://www.w3.org/2003/01/geo/wgs84_pos#",
+    "ldp" : "http://www.w3.org/ns/ldp#",
+    "memento" : "http://mementoweb.org/ns#",
+    "owl" : "http://www.w3.org/2002/07/owl#",
+    "prov" : "http://www.w3.org/ns/prov#",
+    "rdf" : "http://www.w3.org/1999/02/22-rdf-syntax-ns#",
+    "rdfs" : "http://www.w3.org/2000/01/rdf-schema#",
+    "schema" : "http://schema.org/",
+    "skos" : "http://www.w3.org/2004/02/skos/core#",
+    "time" : "http://www.w3.org/2006/time#" ,
+    "xsd" : "http://www.w3.org/2001/XMLSchema#"
+}
+
