+++ trellis-test/src/main/resources/directContainerIsPartOf.ttl
@@ -1,0 +1,8 @@
+PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
+PREFIX ldp: <http://www.w3.org/ns/ldp#>
+PREFIX foaf: <http://xmlns.com/foaf/0.1/>
+PREFIX dc: <http://purl.org/dc/terms/>
+
+<> skos:prefLabel "Direct Container"@eng ;
+   ldp:hasMemberRelation dc:isPartOf ;
+   dc:description "This is a Direct Container for testing."@eng .
